feat: rename useHubSearchPreviewQuery hook and refactor related components for improved preview functionality
